refactor: 엔드포인트 복수형으로 변경
