+++ src/view/group/dashboard/Ranking/TopRanking.tsx
@@ -1,4 +1,4 @@
-import type { RankingResponse } from "@/api/group/type";
+import type { RankingResponse } from "@/api/groups/type";
 import RankingCard from "@/view/group/dashboard/Ranking/RankingCard";
 import { topRankingWrapper } from "@/view/group/dashboard/Ranking/index.css";
